test/fixtures/wpt/webmessaging/broadcastchannel/sandbox.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Creating BroadcastChannel in an opaque origin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
async_test(t => {
    self.onmessage = t.step_func(e => {
        assert_equals(e.data, 'Created');
        t.done();
      });
  });
</script>
<iframe sandbox="allow-scripts" src="resources/sandboxed.html"></iframe>
</body>
