third_party/WebKit/LayoutTests/platform/win/external/wpt/webmessaging/broadcastchannel/sandbox-expected.txt

This is a testharness.js-based test.
FAIL Creating BroadcastChannel in an opaque origin assert_equals: expected "Created" but got "Exception: NotSupportedError"
Harness: the test ran to completion.
